♻️ refactor: stock index socket 디렉토리 위치 이동
